refactored integration routines
