+++ pyerrors/__init__.py
@@ -485,5 +485,6 @@ from . import input
 from . import linalg
 from . import mpm
 from . import roots
+from . import integrate
 
 from .version import __version__
